Workaround Gradle issue with `Configuration`s created via rules (#31)
`Configuration` instances created by rules do not appear to have their
"consumable" flags set properly, leading to failure in the consuming
project.

This change allows the plugin to proactively create the configurations
in the producing module by parsing a gradle property.  This allows us
to work around the issue while the Gradle bug gets raised/fixed/etc.

Once the configurations were explicitly declared, this led to a
disambiguation issue, resolved by adding an attribute to the
configurations to distinguish them.